FIX: Mixed-precision dtype handling for spectral attention
Based on comprehensive research findings:

1. spectral_cache.py:
   - Convert coeffs_q back to self.dtype after INT8 simulation
   - SVD computes in FP32 but downstream ops need matching dtype
   - Fixed both _compress_tensor and _compress_tensor_fallback

2. spectral_attention.py:
   - Use FP32 for einsum (NOT on autocast whitelist)
   - Explicit FP32 cast for Q_aligned, basis_K, coeffs_K before einsum
   - Convert back to model_dtype after computation
   - Same fix applied to value aggregation (attn @ coeffs_V @ basis_V)

3. kernels/spectral_attention.py:
   - spectral_score_forward_pytorch: FP32 einsum with dtype conversion
   - spectral_value_forward_pytorch: FP32 einsum with dtype conversion

4. COLAB_TRITON_TEST.py:
   - Remove .squeeze(2) bug - function expects 4D [B, H_q, 1, T_block]

Research reference: einsum NOT on autocast whitelist, requires explicit dtype match
to avoid FP16 overflow and mixed-precision errors.

+++ unsloth_spectral/spectral_cache.py
@@ -306,7 +306,11 @@ class SpectralCache:
         # Batched quantization (vectorized per-head)
         coeffs_q, scales = self._quantize_int8_batched(coeffs)
         
-        # Keep basis in FP16 (it's small, no need to quantize)
+        # CRITICAL FIX: Convert ALL components back to model dtype (BF16/FP16)
+        # SVD computes in FP32 for numerical stability, but we need matching
+        # dtypes for downstream operations (einsum, matmul with model tensors)
+        # Research finding: einsum NOT on autocast whitelist, requires explicit dtype match
+        coeffs_q = coeffs_q.to(self.dtype)
         basis = basis.to(self.dtype)
         
         return coeffs_q, basis, scales
@@ -343,7 +347,8 @@ class SpectralCache:
             coeffs_h = U_k * S_k.unsqueeze(0)
             coeffs_q, scale = self._quantize_int8(coeffs_h)
             
-            coeffs_list.append(coeffs_q)
+            # CRITICAL FIX: Convert coeffs to model dtype (matching basis)
+            coeffs_list.append(coeffs_q.to(self.dtype))
             basis_list.append(Vh_k.to(self.dtype))
             scales_list.append(scale)
